backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Artem Isachenkov <artem.isachenkov@ofirjob.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.4"
python-dotenv = "^1.0.1"
supabase = "^2.10.0"
sqlmodel = "^0.0.22"
python-jose = "^3.3.0"
alembic = "^1.14.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
